requirements.txt
boto3==1.9.6
Django==2.2.3
django-cors-headers==2.4.0
django-extensions==2.1.2
django-filter==2.0.0
django-graphql-geojson==0.1.4
django-heroku==0.3.1
django-map-widgets==0.2.0
django-storages==1.7.1
geocoder==1.38.1
graphene==2.1.3
graphene-django==2.3.2
graphql-core==2.1
Pillow==5.2.0
psycopg2==2.7.5 --no-binary :all:
requests==2.18.4
urllib3==1.23
uWSGI==2.0.17.1
